Todos los servicios OK

+++ src/main/java/com/uniminuto/clinica/repository/CitaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CitaRepository extends JpaRepository<Cita, Long> { 
+public interface CitaRepository extends JpaRepository<Cita, Integer> { 
 
     @Query("SELECT c FROM Cita c ORDER BY c.fechaHora DESC")
     List<Cita> findAllOrderByFechaHoraDesc();
